refactor form naming GH-88
